+++ src/main/java/com/robosh/service/CarService.java
@@ -12,7 +12,7 @@ public class CarService {
     private DaoFactory daoFactory = DaoFactory.getInstance();
     private static final Logger LOG = Logger.getLogger(CarService.class);
 
-    public Car getCarById(long id){
+    public Car getCarById(int id){
         try (CarDao dao = daoFactory.createCarDao()) {
             LOG.debug("created CarDao");
             return dao.getById(id);
