+++ src/main/java/com/robosh/service/OrderService.java
@@ -12,14 +12,14 @@ public class OrderService {
     private DaoFactory daoFactory = DaoFactory.getInstance();
     private static final Logger LOG = Logger.getLogger(OrderService.class);
 
-    public List<Order> getAllOrderByIdDriver(long idDriver){
+    public List<Order> getAllOrderByIdDriver(int idDriver){
         try (OrderDao dao = daoFactory.createOrderDao()) {
             LOG.debug("created OrderDao");
             return dao.getAllOrdersByDriverId(idDriver);
         }
     }
 
-    public List<Order> getAllOrderByIdDriver(long idDriver, int row, int limit){
+    public List<Order> getAllOrderByIdDriver(int idDriver, int row, int limit){
         try (OrderDao dao = daoFactory.createOrderDao()) {
             LOG.debug("created OrderDao");
             return dao.getAllOrdersByDriverId(idDriver,row, limit);
@@ -32,7 +32,7 @@ public class OrderService {
             return dao.findAll();
         }
     }
-    public Order getOrderById(long idOrder) {
+    public Order getOrderById(int idOrder) {
         try(OrderDao dao = daoFactory.createOrderDao()){
             LOG.debug("created OrderDao");
             return dao.getById(idOrder);
@@ -66,7 +66,7 @@ public class OrderService {
 
     }
 
-    public long getAllOrdersCount(long idDriver){
+    public long getAllOrdersCount(int idDriver){
         try (OrderDao dao = daoFactory.createOrderDao()){
             return dao.getCountOrders(idDriver);
         }
